requirements.txt
# BMAD Creator Management Platform

# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-multipart>=0.0.6

# Auth & Security
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# HTTP Client (for AI providers & social APIs)
httpx>=0.25.0

# Configuration
PyYAML>=6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging
loguru>=0.7.0

# Optional: Redis for rate limiting & caching
# redis>=5.0.0
